Разбивает код на модули(1)
