+++ index.html
@@ -166,6 +166,9 @@
   </template>
 
 
-  <script src="/js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
+  <script src="js/data.js" type="module"></script>
+  <script src="js/util.js" type="module"></script>
+  <script src="js/create-object.js" type="module"></script>
 </body>
 </html>
