+++ js/data.js
@@ -1,0 +1,34 @@
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const MIN_COMMENTS = 0;
+const MAX_COMMENTS = 200;
+const DESC_PHOTOS = [
+  'Пейзаж',
+  'Кот',
+  'Кекс',
+  'Машина',
+  'Дом',
+  'Портрет Кекса',
+  'Велосипед',
+  'Горы',
+  'Город',
+  'Природа',
+  'Водопад',
+  'Дорога',
+  'Лес',
+  'Опушка',
+  'Поляна',
+  'Пикник',
+  'Елка',
+  'Компьютер',
+  'Офис',
+  'Мотоцикл',
+  'Магазин',
+  'Моё',
+  'Твое',
+  'Наше',
+  'Ваше'
+];
+
+export {MIN_LIKES, MAX_LIKES, MIN_COMMENTS, MAX_COMMENTS, DESC_PHOTOS};
+
